File restructuring to clean things up, sockets model
