+++ server/controllers/room.controller.ts
@@ -1,12 +1,37 @@
 import type { OpineRequest, OpineResponse } from "opine";
 import type { RoomSchema } from "../types/schemas.ts";
-import generateRoomCode from "../utils/generateRoomCode.ts";
 import * as rooms from "../models/rooms.ts";
 
 type OpineController<ResBody = unknown> = (
 	req: OpineRequest,
 	res: OpineResponse,
 ) => Promise<OpineResponse<ResBody>>;
+
+const generateRoomCode = async (): Promise<string> => {
+	let roomCode = getRandomCode();
+
+	for (let i = 0; i < 5; i++) {
+		const duplicateCode = await rooms.findByRoomCode(roomCode);
+
+		if (!duplicateCode) return roomCode;
+
+		console.debug(
+			`Duplicate room code found, generating again. Retried ${i + 1} times.`,
+		);
+		roomCode = getRandomCode();
+	}
+
+	throw new Error("Unable to generate unique room code");
+};
+
+const getRandomCode = () => {
+	let text = "";
+	const possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
+	for (let i = 0; i < 4; i++) {
+		text += possible.charAt(Math.floor(Math.random() * possible.length));
+	}
+	return text;
+};
 
 export interface CreateRoomRequest {
 	moderatorName: string;
@@ -43,20 +68,6 @@ export const createRoom: OpineController = async (req, res) => {
 			message: err.message,
 		});
 	}
-};
-
-export const getRoom: OpineController = async (req, res) => {
-	const room = await rooms.findByRoomCode(req.params.roomCode);
-	if (room) {
-		return res.setStatus(200).json({
-			success: true,
-			room,
-		});
-	}
-	return res.setStatus(404).json({
-		success: false,
-		message: `Room with roomCode of ${req.params.roomCode} not found`,
-	});
 };
 
 export const checkRoomExists: OpineController = async (req, res) => {
