+++ server/events/middlewares/mod.ts
@@ -1,0 +1,3 @@
+export { default as validateModerator } from "./validate-moderator.ts";
+export { default as validateVoter } from "./validate-voter.ts";
+export { default as validateInRoom } from "./validate-in-room.ts";
