+++ server/events/start-voting.ts
@@ -1,0 +1,27 @@
+import * as rooms from "../models/rooms.ts";
+import type { EventFunction, StartVotingEvent } from "../types/socket.ts";
+import { sendRoomData } from "./utils/mod.ts";
+
+/**
+ * Begins voting, clearing previous votes
+ */
+const handleStartVoting: EventFunction<StartVotingEvent> = async (roomData) => {
+	const updatedVoters = roomData.voters.map((voter) => ({
+		id: voter.id,
+		name: voter.name,
+		selection: null,
+		confidence: null,
+	}));
+
+	const updatedRoomData = await rooms.updateById(roomData._id, {
+		$set: {
+			state: "Voting",
+			voters: updatedVoters,
+			votingStartedAt: new Date(),
+		},
+	});
+
+	sendRoomData(updatedRoomData);
+};
+
+export default handleStartVoting;
