+++ server/events/utils/mod.ts
@@ -1,0 +1,2 @@
+export { default as cleanseName } from "./cleanse-name.ts";
+export { default as sendRoomData } from "./send-room-data.ts";
