+++ server/events/voter-kicked.ts
@@ -1,0 +1,27 @@
+import * as rooms from "../models/rooms.ts";
+import sockets from "../models/sockets.ts";
+import type { EventFunction, KickVoterEvent } from "../types/socket.ts";
+import { sendRoomData } from "./utils/mod.ts";
+
+const handleVoterKicked: EventFunction<KickVoterEvent> = async (
+	roomData,
+	_,
+	{ voterId },
+) => {
+	const updatedRoomData = await rooms.updateById(roomData._id, {
+		$pull: {
+			voters: {
+				id: voterId,
+			},
+		},
+	});
+
+	if (!updatedRoomData) return;
+
+	sendRoomData(updatedRoomData);
+
+	const kickedVoterSocket = sockets.get(voterId, roomData.roomCode);
+	kickedVoterSocket?.send({ event: "Kicked" });
+};
+
+export default handleVoterKicked;
