--- server/postman/devs-playing-poker local.postman_environment.json
@@ -1,21 +1,0 @@
-{
-	"id": "8c8e0fa7-ff44-49bc-8b8e-9ec7fd193d53",
-	"name": "devs-playing-poker local",
-	"values": [
-		{
-			"key": "PORT",
-			"value": "3000",
-			"type": "default",
-			"enabled": true
-		},
-		{
-			"key": "roomCode",
-			"value": "AAAA",
-			"type": "default",
-			"enabled": true
-		}
-	],
-	"_postman_variable_scope": "environment",
-	"_postman_exported_at": "2022-07-14T18:02:42.330Z",
-	"_postman_exported_using": "Postman/9.24.1"
-}
