+++ server/routes/socket.routes.ts
@@ -1,16 +1,13 @@
 import { Router } from "opine";
-import { handleWs } from "../controllers/socket.controller.ts";
 import { validateOrigin } from "../middlewares/validators.ts";
+import sockets from "../models/sockets.ts";
 
 const router = Router();
 
-router.get("/:roomCode", validateOrigin, async (req, res, next) => {
+router.get("/:roomCode", validateOrigin, (req, res, next) => {
 	if (req.headers.get("upgrade") === "websocket") {
 		const sock = req.upgrade();
-		const userId = req.query.userId ?? crypto.randomUUID();
-		const roomCode = req.params.roomCode;
-
-		await handleWs(sock, userId, roomCode);
+		sockets.add(sock, req.params.roomCode, req.query.userId);
 	} else {
 		return res.send("You've gotta set the magic header...");
 	}
