+++ server/types/socket.ts
@@ -78,3 +78,17 @@ export type WebScoketMessageEvent =
 	| ChangeNameEvent;
 
 export type WebSocketEvent = WebSocketTriggeredEvent | WebScoketMessageEvent;
+
+interface EventFunctionError {
+	message: string;
+}
+
+interface SocketContext {
+	userId: string;
+}
+
+export type EventFunction<Event extends WebScoketMessageEvent> = (
+	roomData: RoomSchema,
+	context: SocketContext,
+	event: Event,
+) => Promise<EventFunctionError | void> | (EventFunctionError | void);
